compiled and tests passed

+++ core/src/main/scala/ru/oseval/datahub/Repeater.scala
@@ -5,7 +5,7 @@ import org.slf4j.Logger
 import scala.concurrent.{ExecutionContext, Future, Promise}
 import scala.util.{Failure, Success}
 
-case class RepeaterConfig(base: Double, maxDelay: Long)
+case class RepeaterConfig(base: Long, maxDelay: Long)
 
 case class Repeater(name: String, config: RepeaterConfig, scheduleOnce: (Long, () => Any) => Unit, log: Logger) {
   def run[T](fut: () => Future[T], attempt: Int = 0, p: Promise[T] = Promise[T])
@@ -14,7 +14,7 @@ case class Repeater(name: String, config: RepeaterConfig, scheduleOnce: (Long, (
       case Success(r) => p.success(r)
       case Failure(e) =>
         val nextAttempt = attempt + 1
-        val delay = Math.pow(config.base, nextAttempt.toDouble) min config.maxDelay
+        val delay = Math.pow(config.base.toDouble, nextAttempt.toDouble) min config.maxDelay
         log.warn(
           "Future inside repeater {} failed with {} and will be recovered with attempt is {} and delay is {}",
           name, e.getMessage, nextAttempt.toString, delay.toString
